chore: ViewModeContainer 공용 컴포넌트화
