+++ client/src/common/ViewModeContainer/BottomProfileBox.styles.ts
