+++ client/src/common/ViewModeContainer/BottomProfileBox.tsx
